Removed RabbitMQ related classes to separate amqp project
